divided socket code into different handlers based on function
