+++ packages/server/src/handlers/GameHandler.ts
@@ -1,0 +1,4 @@
+import { ClientEvents, ServerEvents } from "@hangmen/shared/types/SocketEvent";
+import { Socket, Server } from "socket.io";
+
+export const gameHandler = (io: Server, socket: Socket) => {};
