+++ packages/server/src/server.ts
@@ -5,14 +5,16 @@ import { Server } from "socket.io";
 import cors from "cors";
 import "dotenv/config";
 import {
-  ClientEvents,
   ClientPayloads,
-  ServerEvents,
   ServerPayloads,
 } from "@hangmen/shared/types/SocketEvent";
 
 import { Room } from "./classes/Room";
 import { User } from "./classes/User";
+
+import { userHandler } from "./handlers/UserHandler";
+import { roomHandler } from "./handlers/RoomHandler";
+import { gameHandler } from "./handlers/GameHandler";
 const app = express();
 
 app.use(cors());
@@ -32,78 +34,12 @@ const io = new Server<ClientPayloads, ServerPayloads>(server, {
   },
 });
 
-interface Users {
-  [key: string]: User;
-}
-interface ActiveRooms {
-  [key: string]: Room;
-}
-const users: Users = {};
-const activeRooms: ActiveRooms = {};
-
 io.on("connection", (socket) => {
+  userHandler(io, socket);
+  roomHandler(io, socket);
+  gameHandler(io, socket);
   console.log(`User Connected: ${socket.id}`);
-  socket.on(ClientEvents.REGISTER, (username, userId) => {
-    users[socket.id] = new User(username, userId); //TODO: on new user creation
-  });
 
-  socket.on(ClientEvents.REQUEST_ROOM_CODE, () => {
-    const room = new Room();
-    activeRooms[room.code] = room;
-    socket.emit(ServerEvents.CREATE_ROOM, room.code);
-  });
-  socket.on(ClientEvents.JOIN_ROOM, (roomCode, callback) => {
-    let status = false;
-    if (activeRooms[roomCode]) {
-      socket.join(roomCode);
-      activeRooms[roomCode].addPlayer(socket.id, users[socket.id]);
-      users[socket.id].currentRoom = roomCode;
-      io.to(roomCode).emit(
-        ServerEvents.NEW_PLAYER,
-        users[socket.id],
-        activeRooms[roomCode].getPlayers()
-      );
-      status = true;
-      console.log(activeRooms[roomCode]);
-      console.log(socket.rooms);
-    } else {
-      status = false;
-    }
-    callback({ status });
-  });
-  socket.on(ClientEvents.LEAVE_ROOM, (roomCode, callback) => {
-    //on leave room remove the player from the room, remove the room from the players current room and delete the room if the player count is 0.
-    if (roomCode in activeRooms && socket.id in activeRooms[roomCode].players) {
-      delete activeRooms[roomCode].players[socket.id];
-      users[socket.id].currentRoom = "";
-      if (Object.keys(activeRooms[roomCode].players).length === 0) {
-        delete activeRooms[roomCode];
-      }
-      socket.leave(roomCode);
-      if (
-        !(roomCode in activeRooms && socket.id in activeRooms[roomCode].players)
-      ) {
-        callback({
-          status: "ok",
-        });
-      }
-      io.to(roomCode).emit(ServerEvents.PLAYER_LEAVE_ROOM, users[socket.id]);
-    } else {
-      callback({
-        status: "error",
-        message: "Error leaving the room. Room or player not found.",
-      });
-    }
-  });
-  socket.on(ClientEvents.REJOIN_ROOM, (roomCode, cb) => {});
-  socket.on(ClientEvents.SEND_MESSAGE, (id, text) => {
-    //this works fine for now but I need to add a timestamp and determine if to keep io functionality or not
-    io.to(users[socket.id].currentRoom).emit(ServerEvents.RECEIVE_MESSAGE, {
-      id,
-      sender: users[socket.id].username,
-      text,
-    });
-  });
   socket.on("disconnect", () => {
     console.log("User Disconnected:", socket.id);
   });
